Dev 1

+++ app/model/staff_model.php
@@ -23,7 +23,7 @@ class staff_model extends model {
     }
 
     function delete($id) {
-        $sth = $this->db->prepare("DELETE FROM staff WHERE id = $id;");
+        $sth = $this->db->prepare("DELETE FROM staff WHERE id = $id AND `admin` = 0;");
         $sth->execute();
     }
 
@@ -33,7 +33,15 @@ class staff_model extends model {
     }
 
     function create($data) {
-        return;
+        $sth = $this->db->prepare("INSERT INTO staff(username,email,passwordhash,`admin`) VALUES(?,?,?,?);");
+        try{
+            $sth->execute([$data["username"], $data["email"], $data["passwordhash"], $data["admin"]]);
+        }
+        catch (PDOException $e)
+        {
+            return false;
+        }
+        return true;
     }
 
 }
